// src/types/allegro.ts
export interface AllegroDispute {
  id: string;
  subject: {
    id: string;
    name: string;
  };
  status: 'ONGOING' | 'CLOSED' | 'UNRESOLVED';
  messagesStatus: 'NEW' | 'BUYER_REPLIED' | 'SELLER_REPLIED';
  buyer: {
    id: string;
    login: string;
  };
  checkoutForm: {
    id: string;
    createdAt: string;
  };
  createdAt: string;
  updatedAt: string;
}

export interface AllegroDisputeMessage {
  id: string;
  text: string;
  type: 'REGULAR' | 'END_REQUEST';
  author: {
    login: string;
    role: 'BUYER' | 'SELLER' | 'ALLEGRO_ADVISOR';
  };
  attachment?: {
    fileName: string;
    url: string;
  };
  createdAt: string;
}